deleteRestaurant

+++ restaurants/lambda/dbaccess/dbaccess/methods/deleteRestaurant.js
@@ -9,32 +9,46 @@ var deleteReviewForRestaurant = require('./deleteReviewForRestaurant');
  */
 deleteRestaurant = async function(client, restaurantId) {
 
-
+    var success = { statusCode: 200, body: null };
+    var notFound = { statusCode: 404, body: null };
     var runIt = async function(resolve, reject) {
 
         var foundRestaurant;
 
         try {
             foundRestaurant = await getSingleRestaurant(client, restaurantId);
+
             // console.log("found " + foundRestaurant.id)
-            foundRestaurant.reviewDTOs.forEach(rev => {
+            if (foundRestaurant) {
+                foundRestaurant.reviewDTOs.forEach(rev => {
 
-                //TODO delete the reviews
-                //  await deleteReviewForRestaurant(restaurantId, rev.id);
+                    //TODO delete the reviews
+                    deleteReviewForRestaurant(client, restaurantId, rev.id).then((d) => {
+                        // console.log(d + "xxxxx")
+                    });
 
-            });
-            var params = { TableName: "Restaurants", Key: { id: id } }
-            client.delete(params, (err, data) => {
+                });
 
-                if (err) reject(err); // an error occurred
-                else resolve(null); // successful response
+                var params = {
+                    TableName: "Restaurants",
+                    Key: {
+                        id: restaurantId
+                    }
+                }
+                client.delete(params, (err, data) => {
 
-            })
+                    if (err) reject(err); // an error occurred
+                    else resolve(success); // successful response
 
+                })
 
+            } else {
+                console.log(`restaurant ${restaurantId}  not found`);
+                resolve(notFound)
+            }
 
         } catch (ee) {
-            reject(ee)
+            reject(ee);
         }
 
     }
